backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pymysql==1.1.1
pydantic-settings==2.8.1
python-jose==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==44.0.2
